requirements.txt
torch
torchvision
tqdm
numpy
opencv-python
mediapipe
pillow
streamlit
streamlit-webrtc
av
pyautogui 